+++ bricks/test_optimizer/hooks/analysis_options.yaml
@@ -1,1 +1,4 @@
-include: package:very_good_analysis/analysis_options.3.1.0.yaml
+include: package:very_good_analysis/analysis_options.4.0.0.yaml
+linter:
+  rules:
+    public_member_api_docs: false
