+++ bricks/test_optimizer/hooks/pubspec.yaml
@@ -6,10 +6,9 @@ environment:
 
 dependencies:
   mason: ">=0.1.0-dev.41 <0.1.0"
-  test: ^1.22.2
   path: ^1.8.1
 
 dev_dependencies:
-  checked_yaml: ^2.0.2
   mocktail: ^0.3.0
-  very_good_analysis: ^3.1.0
+  test: ^1.22.2
+  very_good_analysis: ^4.0.0
