--- bricks/test_optimizer/hooks/test/test_dir/test/some_file.dart
@@ -1,1 +1,0 @@
-
