--- bricks/test_optimizer/hooks/test/test_dir/test/some_test.dart
@@ -1,1 +1,0 @@
-
