+++ tool/generate_test_optimizer_bundle.sh
@@ -1,5 +1,4 @@
 #!/bin/bash
-# Runs `mason bundle` to generate the bundle for the optimizer used in the test command
 
 mason bundle --source path ./bricks/test_optimizer/ -t dart --output-dir lib/src/commands/test/templates
